Move alert to shared views.
