+++ contentcuration/contentcuration/frontend/channelEdit/views/files/thumbnails/ThumbnailGenerator.vue
@@ -19,7 +19,7 @@
   import map from 'lodash/map';
   import max from 'lodash/max';
   import pdfJSLib from 'pdfjs-dist';
-  import Alert from 'edit_channel/sharedComponents/Alert.vue';
+  import Alert from 'shared/views/Alert';
   import Uploader from 'frontend/channelEdit/views/files/Uploader';
 
   const MAX_AUDIO_SAMPLE_SIZE = 64000;
